# Add benchmark for powershell parser and extend tests