--- a/.golangci.yml
+++ b/.golangci.yml
@@ -1,5 +1,7 @@
 run:
-  deadline: 1m
+  timeout: 5m
+  skip-files:
+    - '(.+)_test\.go'
 
 linters:
   disable-all: false