--- a/powershell_test.go
+++ b/powershell_test.go
@@ -6,6 +6,31 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+func BenchmarkGetPowershellArgs(b *testing.B) {
+	b.ReportAllocs()
+
+	for i := 0; i < b.N; i++ {
+		GetPowershellArgs([]string{
+			"--powershell-api",
+			"https://battlestation:5668",
+			"--powershell-insecure",
+			"-C",
+			"try { Use-Icinga -Minimal; } catch { <# some error #>; exit 3; }; " +
+				"Exit-IcingaExecutePlugin -Command 'Invoke-IcingaCheckUsedPartitionSpace' ",
+			"-Warning",
+			"80",
+			"-Critical",
+			"95",
+			"-Include",
+			"@()",
+			"-Exclude",
+			"@('abc','def')",
+			"-Verbosity",
+			"2",
+		})
+	}
+}
+
 func TestGetPowershellArgs(t *testing.T) {
 	command, args := GetPowershellArgs([]string{"-C", "Invoke-IcingaCheckUsedPartitionSpace", "-Warning", "80"})
 	assert.Equal(t, "Invoke-IcingaCheckUsedPartitionSpace", command)