--- a/testdata/netstring.dat
+++ b/testdata/netstring.dat
@@ -1,1 +1,0 @@
-1:a,5:bbbbb,11:aaaaaaaaaaa,