Tried to optimize some functions

Before:
```
pkg: github.com/NETWAYS/icinga-powershell-connector
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
BenchmarkGetPowershellArgs-8   	  444860	      2591 ns/op	     988 B/op	      22 allocs/op
BenchmarkGetPowershellArgs-8   	  479780	      2557 ns/op	     988 B/op	      22 allocs/op
BenchmarkGetPowershellArgs-8   	  423138	      2580 ns/op	     988 B/op	      22 allocs/op
```

After:
```
pkg: github.com/NETWAYS/icinga-powershell-connector
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
BenchmarkGetPowershellArgs-8   	  968371	      1107 ns/op	     872 B/op	      19 allocs/op
BenchmarkGetPowershellArgs-8   	  967675	      1096 ns/op	     872 B/op	      19 allocs/op
BenchmarkGetPowershellArgs-8   	  955102	      1091 ns/op	     872 B/op	      19 allocs/op

GetPowershellArgs-8        2.586µ ± 1%   1.106µ ± 1%  -57.22% (p=0.000 n=12)
```