@martialblog 2023-03-06
@RincewindsHat This would not work with the new version, should this work?

```
command = ParsePowershellTryCatch("Exit-Foo -Test \"'Invoke-IcingaCheckUsedPartitionSpace'\"")
assert.Equal(t, "Invoke-IcingaCheckUsedPartitionSpace", command)
```