@martialblog 2023-03-06
@RincewindsHat I extended the tests and updated the function. I ran the new tests both on `main` and on this branch.