@martialblog 2023-03-07
@RincewindsHat I abused this PR to add some additional tests... :shipit: